+++ features/Context/Loader/UserLoader.php
@@ -4,7 +4,6 @@ namespace Context\Loader;
 
 use Oro\Bundle\UserBundle\Entity\Group;
 use Symfony\Component\Yaml\Yaml;
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Persistence\ObjectManager;
 use Oro\Bundle\UserBundle\Entity\Role;
 use Oro\Bundle\UserBundle\Entity\UserApi;
@@ -81,19 +80,13 @@ class UserLoader extends LoadUserData
         $api = new UserApi();
         $api->setApiKey($apiKey)->setUser($user);
 
-        $unit = $this->getOwner('Main');
-
         $user
             ->setUsername($username)
             ->setPlainPassword($password)
             ->setFirstname($firstName)
             ->setLastname($lastName)
             ->setEmail($email)
-            ->setApi($api)
-            ->setOwner($unit)
-            ->setBusinessUnits(
-                new ArrayCollection(array($unit))
-            );
+            ->setApi($api);
 
         foreach ($roles as $role) {
             $user->addRole($this->getOrCreateRole($role));
@@ -111,7 +104,6 @@ class UserLoader extends LoadUserData
         // Following to fix a cascade persist issue on UserApi occuring only during Behat Execution
         $this->getUserManager()->getStorageManager()->clear('Oro\Bundle\UserBundle\Entity\User');
         $this->getUserManager()->getStorageManager()->clear('Oro\Bundle\UserBundle\Entity\UserApi');
-        $this->getUserManager()->getStorageManager()->clear('OroEmail\Cache\OroEmailBundle\Entity\EmailAddressProxy');
     }
 
     /**