+++ src/Pim/Bundle/VersioningBundle/Builder/VersionBuilder.php
@@ -47,7 +47,7 @@ class VersionBuilder
         $oldSnapshot   = $previousVersion ? $previousVersion->getSnapshot() : [];
 
         // TODO: we don't use direct json serialize due to convert to audit data based on array_diff
-        $snapshot = $this->normalizer->normalize($versionable, 'csv', array('versioning' => true));
+        $snapshot = $this->normalizer->normalize($versionable, 'csv', ['versioning' => true]);
 
         $changeset = $this->buildChangeset($oldSnapshot, $snapshot);